move unit tests to sub folder
