+++ package.json
@@ -31,9 +31,9 @@
     "test-lib-view": "NODE_ENV=test-lib yarn run start",
     "test-loki-view": "NODE_ENV=test-loki yarn run start",
     "test-electron": "yarn grunt test",
-    "test-integration": "ELECTRON_DISABLE_SANDBOX=1 mocha --exit --timeout 10000 ts/test/session/integration/integration_test.js",
+    "test-integration": "ELECTRON_DISABLE_SANDBOX=1 mocha --exit --full-trace --timeout 10000 ts/test/session/integration/integration_test.js",
     "test-node": "mocha --recursive --exit test/app test/modules ts/test libloki/test/node --timeout 10000",
-    "test-session": "mocha --recursive --exit ts/test/session --timeout 10000",
+    "test-session": "mocha --recursive --exit ts/test/session/unit --full-trace --timeout 10000",
     "test-medium-groups": "ELECTRON_DISABLE_SANDBOX=1 mocha --exit --timeout 10000 integration_test/integration_test.js --grep 'senderkeys'",
     "test-groups": "ELECTRON_DISABLE_SANDBOX=1 mocha --exit --timeout 10000 integration_test/integration_test.js --grep 'Closed groups'",
     "eslint": "eslint --cache .",
