+++ ts/test/session/unit/sending/PendingMessageCache_test.ts
@@ -1,10 +1,10 @@
 import { expect } from 'chai';
 import * as sinon from 'sinon';
 import * as _ from 'lodash';
-import { MessageUtils } from '../../../session/utils';
-import { TestUtils } from '../../../test/test-utils';
-import { PendingMessageCache } from '../../../session/sending/PendingMessageCache';
-import { SessionProtocol } from '../../../session/protocols';
+import { MessageUtils } from '../../../../session/utils';
+import { TestUtils } from '../../../../test/test-utils';
+import { PendingMessageCache } from '../../../../session/sending/PendingMessageCache';
+import { SessionProtocol } from '../../../../session/protocols';
 
 // Equivalent to Data.StorageItem
 interface StorageItem {
